# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Knock-knock REST API",
        "title": "Knock-knock REST API",
        "contact": {
            "name": "API Support",
            "url": "http://AN_ORG.github.io",
            "email": "AN_EMAIL"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/knock-knock",
    "paths": {
...
        "/tb/auth/test": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Auth test with TB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Auth] Test with TB"
                ],
                "summary": "Auth test with TB.",
                "responses": {
                    "200": {
                        "description": "Auth info for test",
                        "schema": {
                            "$ref": "#/definitions/controller.AuthsInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.AuthsInfo": {
            "type": "object",
            "properties": {
                "authenticated": {
                    "type": "boolean"
                },
                "expired-time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "model.BasicResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token (get token in http://localhost:8056/auth)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}